tutorial/requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
httpie==2.1.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
psycopg2==2.8.5
Pygments==2.6.1
pylint==2.5.2
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
